requirements.txt
flask
python-dotenv
psycopg
redis